backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
streamlit
python-multipart
email-validator